criação de alguns endpoints

+++ grupo5/src/main/java/br/com/devinhouse/grupo5/dto/ProcessoOutputDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString
 public class ProcessoOutputDTO {
 
-    private Long nuProcesso;
+    private Integer nuProcesso;
     private String sgOrgaoProcesso;
     private String nuAnoProcesso;
     private String descricao;
